refact: learning about `modularization`
  i refact my project to separate `modules` and `bounded contexts`
