+++ src/domain/checkout/repository/order-repository.interface.ts
@@ -1,0 +1,6 @@
+import RepositoryInterface from "../../@shared/repository/repository-interface";
+import Order from "../orders/order";
+
+export default interface IOderRepository extends RepositoryInterface<Order> {} {
+  
+}
