+++ src/domain/customer/entity/customer.spec.ts
