--- src/domain/order/order-repository.interface.ts
@@ -1,6 +1,0 @@
-import RepositoryInterface from "../@shared/repository/repository-interface";
-import Order from "../entity/orders/order";
-
-export default interface IOderRepository extends RepositoryInterface<Order> {} {
-  
-}
