+++ src/domain/product/entity/product.interface.ts
