+++ src/domain/product/event/handler/send-email.handler.ts
@@ -1,5 +1,5 @@
-import IEventHandler from "../../event-handler.interface";
-import ProductCreatedEvent from "../product-created.event";
+import IEventHandler from "../../../event/@shared/event-handler.interface";
+import ProductCreatedEvent from "./product-created.event";
 
 export default class SendEmailWhenProductIsCreatedHandler implements IEventHandler<ProductCreatedEvent> {
   handle(event: ProductCreatedEvent): void {
