+++ src/domain/product/repository/product-repository.ts
@@ -1,0 +1,4 @@
+import Product from "../entity/product";
+import IRepository from "./repository-interface";
+
+export default interface ProductInterface extends IRepository<Product> {}
