+++ src/domain/product/repository/repository-interface.ts
