--- src/domain/repository/product/product-repository.ts
@@ -1,4 +1,0 @@
-import Product from "../../entity/product/product";
-import IRepository from "../repository-interface";
-
-export default interface ProductInterface extends IRepository<Product> {}
