+++ src/infra/customer/repository/sequelize/customer.repository.ts
@@ -1,5 +1,5 @@
+import Customer from "../../../../domain/customer/entity/customer";
 import CustomerRepositoryInterface from "../../../../domain/customer/repository/customer-repository.interface";
-import Customer from "../../../../domain/entity/customer/customer";
 import Address from "../../../../value-object/address";
 import CustomerModel from "./customer.model";
