+++ src/infra/product/repository/sequelize/product.repository.spec.ts
@@ -1,5 +1,5 @@
 import { Sequelize } from "sequelize-typescript";
-import Product from "../../../../domain/entity/product/product";
+import Product from "../../../../domain/product/entity/product";
 import ProductModel from "./product.model";
 import ProductRepository from "./product.repository";
